+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ### 2025-06-06
+- [Patch v5.9.2] Add unit tests for order_manager
+- New/Updated unit tests added for tests.test_order_manager_module
+- QA: pytest -q passed (509 tests)
+
+### 2025-06-06
 - [Patch v5.9.1] Validate trade log not empty in real_train_func
 - New/Updated unit tests added for tests.test_training_empty_log::test_real_train_func_empty_trade_log
 - QA: pytest -q passed (505 tests)